refactor: separates `server actions` from `pages` in different files
