+++ src/app/auth/forget-password/page.tsx
@@ -1,65 +1,15 @@
-import { revalidatePath } from 'next/cache';
 import Link from 'next/link';
 
-import { ForgetPasswordEmail } from '@/components/email-templates/ForgetPassword';
 import { SubmitButton } from '@/components/SubmitButton';
 import { Button } from '@/components/ui/Button';
 import { Input } from '@/components/ui/Input';
-import { createAuthenticationDataSource } from '@/data/authentication';
-import {
-  auth,
-  ForgetPasswordInput,
-  ForgetPasswordOutput,
-} from '@/models/authentication';
-import { emailService } from '@/models/email';
-import { feedbackMessage } from '@/src/utils/feedbackMessage';
-import { form } from '@/src/utils/form';
+import { auth } from '@/models/authentication';
 
-let responseMessage: ForgetPasswordOutput;
-let successMessage: {
-  email: string;
-};
-
-async function forgetPassword(formData: FormData) {
-  'use server';
-
-  const { email } = form.sanitizeData<ForgetPasswordInput>(formData);
-
-  const authDataSource = createAuthenticationDataSource();
-  responseMessage = await auth.forgetPassword(authDataSource, {
-    email,
-  });
-
-  if (responseMessage.data) {
-    successMessage = { email };
-    const { resetPasswordTokenId, name } = responseMessage.data;
-
-    await emailService.sendResetPasswordEmail({
-      from: 'Onde Ir <onboarding@resend.dev>',
-      to: email,
-      content: ForgetPasswordEmail({
-        userFirstname: name,
-        resetPasswordTokenId,
-      }),
-    });
-
-    feedbackMessage.setFeedbackMessage({
-      type: 'success',
-      content: 'Instruções enviadas para o email informado',
-    });
-
-    return;
-  }
-
-  revalidatePath('/auth/forget-password');
-
-  feedbackMessage.setFeedbackMessage({
-    type: 'error',
-    content: responseMessage.error.message,
-  });
-}
+import { store } from './store';
 
 export default function Page() {
+  const { responseMessage, successMessage } = store.getForgetPasswordResponse();
+
   return (
     <div className="space-y-4 px-5 md:w-96">
       <h1 className="text-center text-2xl">
@@ -69,7 +19,7 @@ export default function Page() {
         Digite seu email para enviarmos as instruções de recuperação de senha
       </h2>
 
-      <form action={forgetPassword} className="flex flex-col space-y-3">
+      <form action={store.forgetPassword} className="flex flex-col space-y-3">
         <Input
           id="email"
           placeholder="Email*"
