+++ src/app/auth/signin/page.tsx
@@ -1,48 +1,10 @@
-import { cookies } from 'next/headers';
 import Link from 'next/link';
-import { redirect } from 'next/navigation';
 
 import { SubmitButton } from '@/components/SubmitButton';
 import { Input } from '@/components/ui/Input';
-import { createAuthenticationDataSource } from '@/data/authentication';
-import { auth, SignInProps, SignInResponse } from '@/models/authentication';
-import { constants } from '@/src/utils/constants';
-import { feedbackMessage } from '@/src/utils/feedbackMessage';
-import { form } from '@/src/utils/form';
+import { auth } from '@/models/authentication';
 
-let signInResponse: SignInResponse;
-
-async function signInAction(formData: FormData) {
-  'use server';
-
-  const sanitizedData = form.sanitizeData<SignInProps>(formData);
-
-  const authDataSource = createAuthenticationDataSource();
-  signInResponse = await auth.signIn(authDataSource, sanitizedData);
-
-  if (signInResponse.data) {
-    const { accessToken } = signInResponse.data;
-
-    const sevenDaysInMilliseconds = 60 * 60 * 24 * 7 * 1000;
-
-    cookies().set(constants.accessTokenKey, accessToken, {
-      expires: new Date(Date.now() + sevenDaysInMilliseconds),
-      httpOnly: true,
-    });
-
-    feedbackMessage.setFeedbackMessage({
-      type: 'success',
-      content: 'Login efetuado com sucesso!',
-    });
-
-    redirect('/dashboard');
-  }
-
-  feedbackMessage.setFeedbackMessage({
-    type: 'error',
-    content: signInResponse.error.message,
-  });
-}
+import { store } from './store';
 
 type Props = {
   searchParams: {
@@ -51,6 +13,8 @@ type Props = {
 };
 
 export default function Page({ searchParams }: Props) {
+  const { signInResponse } = store.getSignInResponse();
+
   const { userName } = searchParams;
 
   return (
@@ -76,7 +40,7 @@ export default function Page({ searchParams }: Props) {
         </Link>
       </p>
 
-      <form action={signInAction} className="flex flex-col gap-2">
+      <form action={store.signInAction} className="flex flex-col gap-2">
         <Input
           placeholder="Email*"
           id="email"
